final changes

+++ src/services/ApiComentario.ts
@@ -6,15 +6,12 @@ export const getComentarios = async () => {
   return response.data
 }
 
-export const getComentario = async (id:number) => {
-  const response = await api.get(`/comentario/${id}`)
-  return response.data
-}
-
 export const createComentario = async (data : Comentario) => {
-  return await api.post(`/comentario`,data)
+  const token = localStorage.getItem("token");
+  return await api.post(`/comentario`,data,{headers:{Authorization : `Bearer ${token}`}})
 }
 
 export const delComentario = async (id : Number) => {
-  return await api.delete(`/comentario/${id}`)
+  const token = localStorage.getItem("token");
+  return await api.delete(`/comentario/${id}`,{headers:{Authorization : `Bearer ${token}`}})
 }
